Port the original JS repros consolidation from master
