--- e2e/test/scenarios/models/reproductions/20042-nodata-user-blank-screen.cy.spec.js
@@ -1,29 +1,0 @@
-import { ORDERS_QUESTION_ID } from "e2e/support/cypress_sample_instance_data";
-import { restore } from "e2e/support/helpers";
-
-describe("issue 20042", () => {
-  beforeEach(() => {
-    cy.intercept("POST", `/api/card/${ORDERS_QUESTION_ID}/query`).as("query");
-
-    restore();
-    cy.signInAsAdmin();
-
-    cy.request("PUT", `/api/card/${ORDERS_QUESTION_ID}`, {
-      name: "Orders Model",
-      type: "model",
-    });
-
-    cy.signIn("nodata");
-  });
-
-  it("nodata user should not see the blank screen when visiting model (metabase#20042)", () => {
-    cy.visit(`/model/${ORDERS_QUESTION_ID}`);
-
-    cy.wait("@query");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Orders Model");
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.contains("37.65");
-  });
-});
