--- e2e/test/scenarios/models/reproductions/22518.cy.spec.js
@@ -1,46 +1,0 @@
-import {
-  restore,
-  openQuestionActions,
-  summarize,
-  sidebar,
-} from "e2e/support/helpers";
-
-describe("issue 22518", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-    cy.intercept("POST", "/api/dataset").as("dataset");
-    cy.createNativeQuestion(
-      {
-        native: {
-          query: "select 1 id, 'a' foo",
-        },
-        type: "model",
-      },
-      { visitQuestion: true },
-    );
-  });
-
-  it("UI should immediately reflect model query changes upon saving (metabase#22518)", () => {
-    openQuestionActions();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Edit query definition").click();
-
-    cy.get(".ace_content").type(", 'b' bar");
-    cy.findByTestId("native-query-editor-container").icon("play").click();
-    cy.wait("@dataset");
-
-    cy.findByTestId("dataset-edit-bar").button("Save changes").click();
-
-    cy.findAllByTestId("header-cell")
-      .should("have.length", 3)
-      .and("contain", "BAR");
-
-    summarize();
-
-    sidebar()
-      .should("contain", "ID")
-      .and("contain", "FOO")
-      .and("contain", "BAR");
-  });
-});
