--- e2e/test/scenarios/models/reproductions/25537-model-picker-locale.cy.spec.js
@@ -1,40 +1,0 @@
-import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
-import {
-  entityPickerModal,
-  restore,
-  startNewQuestion,
-} from "e2e/support/helpers";
-
-const { ORDERS_ID } = SAMPLE_DATABASE;
-
-const questionDetails = {
-  name: "Orders model",
-  query: { "source-table": ORDERS_ID },
-  type: "model",
-};
-
-describe("issue 25537", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-    cy.intercept("GET", "/api/collection/*/items?*").as("getCollectionContent");
-  });
-
-  it("should be able to pick a saved model when using a non-english locale (metabase#25537)", () => {
-    setLocale("de");
-    cy.createQuestion(questionDetails);
-
-    startNewQuestion();
-    entityPickerModal().within(() => {
-      cy.findByText(/Modelle/i).click();
-      cy.wait("@getCollectionContent");
-      cy.findByText(questionDetails.name).should("exist");
-    });
-  });
-});
-
-const setLocale = locale => {
-  cy.request("GET", "/api/user/current").then(({ body: { id: user_id } }) => {
-    cy.request("PUT", `/api/user/${user_id}`, { locale });
-  });
-};
