--- e2e/test/scenarios/models/reproductions/37009-stale-result-metadata.cy.spec.js
@@ -1,74 +1,0 @@
-import {
-  focusNativeEditor,
-  modal,
-  openQuestionActions,
-  popover,
-  restore,
-} from "e2e/support/helpers";
-
-describe("issue 37009", () => {
-  beforeEach(() => {
-    restore();
-    cy.intercept("POST", "/api/dataset").as("dataset");
-    cy.intercept("POST", "/api/card").as("saveCard");
-    cy.intercept("PUT", "/api/card/*").as("updateCard");
-    cy.signInAsNormalUser();
-  });
-
-  it("should prevent saving new and updating existing models without result_metadata (metabase#37009)", () => {
-    cy.visit("/model/new");
-    cy.findByTestId("new-model-options")
-      .findByText("Use a native query")
-      .click();
-
-    focusNativeEditor().type("select * from products");
-    cy.findByTestId("dataset-edit-bar")
-      .button("Save")
-      .should("be.disabled")
-      .trigger("mousemove", { force: true });
-    cy.findByRole("tooltip").should(
-      "have.text",
-      "You must run the query before you can save this model",
-    );
-    cy.findByTestId("native-query-editor-container").icon("play").click();
-    cy.wait("@dataset");
-    cy.findByRole("tooltip").should("not.exist");
-    cy.findByTestId("dataset-edit-bar")
-      .button("Save")
-      .should("be.enabled")
-      .click();
-    modal()
-      .last()
-      .within(() => {
-        cy.findByLabelText("Name").type("Model");
-        cy.findByText("Save").click();
-      });
-    cy.wait("@saveCard")
-      .its("request.body")
-      .its("result_metadata")
-      .should("not.be.null");
-
-    openQuestionActions();
-    popover().findByText("Edit query definition").click();
-    focusNativeEditor().type(" WHERE CATEGORY = 'Gadget'");
-    cy.findByTestId("dataset-edit-bar")
-      .button("Save changes")
-      .should("be.disabled")
-      .trigger("mousemove", { force: true });
-    cy.findByRole("tooltip").should(
-      "have.text",
-      "You must run the query before you can save this model",
-    );
-    cy.findByTestId("native-query-editor-container").icon("play").click();
-    cy.wait("@dataset");
-    cy.findByRole("tooltip").should("not.exist");
-    cy.findByTestId("dataset-edit-bar")
-      .button("Save changes")
-      .should("be.enabled")
-      .click();
-    cy.wait("@updateCard")
-      .its("request.body")
-      .its("result_metadata")
-      .should("not.be.null");
-  });
-});
